+++ examples/04-frontier-faces/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @cotal-ai/example-04-frontier-faces
 
+## 0.0.30
+
+### Patch Changes
+
+- Updated dependencies [4cf5f72]
+- Updated dependencies [219d33c]
+- Updated dependencies [9c2412c]
+  - @cotal-ai/core@0.21.0
+  - @cotal-ai/workspace@0.21.0
+
+## 0.0.29
+
+### Patch Changes
+
+- Updated dependencies [2752fe7]
+  - @cotal-ai/core@0.20.1
+  - @cotal-ai/workspace@0.20.1
+
 ## 0.0.28
 
 ### Patch Changes