+++ implementations/web/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @cotal-ai/web
 
+## 0.21.0
+
+## 0.20.1
+
 ## 0.20.0
 
 ### Minor Changes